need to get ra and dec delta for each night and append to bigdata

+++ runsextractormasking.py
@@ -122,7 +122,7 @@ def run(imagefilename,weightfilename,survey='DES',index='',bigreturn=False):
 
 
     from scipy import fftpack
-    F1 = fftpack.fft2(im)
+    F1 = fftpack.fft2(im.astype(float))
     F2 = fftpack.fftshift(F1)
     psd2D = np.abs(F2) ** 2
     psd1D = radialProfile.azimuthalAverage(psd2D)
